Move generateUniqueShortId to a new file
